fix: show header row when column has header renderer and null header
A null header should only suppress the default header text, not hide
a row rendered by headerRenderer. Also extract CellContentDirective
into its own module and turn the row visibility checks into pure
functions.
